% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displayText.r
\name{displayText}
\alias{displayText}
\title{displays an object or text to the screen}
\usage{
displayText(id, input = NULL, output = NULL, ui = T, text = "none given")
}
\arguments{
\item{id}{shiny ID}

\item{input}{shiny input object}

\item{output}{shiny output object}

\item{ui}{TRUE selects the UI o/w server side}

\item{text}{reactive list with data and current states}
}
\description{
displays an object or text to the screen
}
