% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explorer.r
\name{explorer}
\alias{explorer}
\title{explores one set of data, e.g. THP-1}
\usage{
explorer(id, input = NULL, output = NULL, session = NULL, ui = T, rv = NULL)
}
\arguments{
\item{id}{shiny ID}

\item{input}{shiny input object}

\item{output}{shiny output object}

\item{ui}{TRUE selects the UI o/w server side}

\item{rv}{reactive list with data and current states}
}
\description{
explores one set of data, e.g. THP-1
}
