% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneCopyTool.r
\name{geneCopyTool}
\alias{geneCopyTool}
\title{widget to copy gene list or details}
\usage{
geneCopyTool(id, input = NULL, output = NULL, ui = T, genes = NULL, rv = NULL)
}
\arguments{
\item{genes}{a reactive function listing the selected genes}

\item{rv}{a reactive object holding the current data}
}
\value{

}
\description{
Allows the user to download either a list (e.g. for enrichment) or
full information of a list of genes.
}
