% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneDistributions.r
\name{geneDistributions}
\alias{geneDistributions}
\title{Title}
\usage{
geneDistributions(id, input = NULL, output = NULL, ui = T, rv = NULL)
}
\arguments{
\item{rv}{}
}
\value{

}
\description{
Title
}
