% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneFilter.r
\name{geneFilter}
\alias{geneFilter}
\title{gene filter widget}
\usage{
geneFilter(
  id,
  input = NULL,
  output = NULL,
  session = NULL,
  ui = T,
  rv = NULL,
  selectionPMA = NULL
)
}
\arguments{
\item{id}{shiny ID}

\item{input}{shiny input object}

\item{output}{shiny output object}

\item{ui}{TRUE selects the UI o/w server side}

\item{rv}{reactive list with data and current states}

\item{selectionPMA}{a reactive list for the selected PMA values}
}
\description{
gene filter widget
}
