% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericDownloadButton.r
\name{genericDownloadButton}
\alias{genericDownloadButton}
\title{a generic data download widget}
\usage{
genericDownloadButton(
  id,
  input = NULL,
  output = NULL,
  ui = T,
  data = NULL,
  label = "download"
)
}
\arguments{
\item{id}{the Shiny ID}

\item{input}{shiny input list}

\item{output}{Shiny output list}

\item{ui}{if TRUE (default) returns the UK}

\item{data}{the data to be downloaded if requested}

\item{label}{a label to be displayed by the download icon}
}
\value{

}
\description{
Am easy to use wrapper for the Shiny download functionality with some
arbitrary defaults.
}
