% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneViewer.r
\name{geneViewer}
\alias{geneViewer}
\title{panel to display user selected genes}
\usage{
geneViewer(id, input = NULL, output = NULL, ui = T, rv = rv)
}
\arguments{
\item{id}{shiny ID}

\item{input}{shiny input object}

\item{output}{shiny output object}

\item{ui}{TRUE selects the UI o/w server side}

\item{rv}{reactive list with data and current states}
}
\description{
Allows users to copy and past lists of gene names into a text box and display the results as a series of profile plots.
}
