% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iwindow.r
\name{iwindow}
\alias{iwindow}
\title{returns a right-handed section from a vector 1:n}
\usage{
iwindow(i, w, n)
}
\arguments{
\item{i}{the index from with the section is to be made}

\item{w}{the width of (number of elements in) the section}

\item{n}{the value of n}
}
\value{
an integer vector
}
\description{
Returns a right-handed section from the vector 1:n if necessary wrapping round. If the section
width equals or is greater than the vector length, n, the entire vector is returned
}
\examples{
iwindow(5,3,10)
iwindow(5,7,10)
iwindow(5,20,10)
iwindow(5,0,10)
}
