% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiGrepl.r
\name{multiGrepl}
\alias{multiGrepl}
\title{a multiple pattern grepl}
\usage{
multiGrepl(patterns, x)
}
\arguments{
\item{patterns}{a string vector of patterns}

\item{x}{a string vector to be matched}
}
\value{
a logical vector of the same length as x
}
\description{
Applies multiple patterns to a string returning a TRUE if any of them
match
}
\examples{

a=c("Aaa","bbC","ee","ff")
multiGrepl(c("aa","ee"),a)
}
