% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newProfiles.r
\docType{data}
\name{newProfiles}
\alias{newProfiles}
\title{data for profile visualisation}
\format{
a standard R list
}
\usage{
data(newProfiles)
}
\description{
Values written by CTTV-020-31-4
}
\keyword{datasets}
