% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseGeneNames.r
\name{parseGeneNames}
\alias{parseGeneNames}
\title{parses out the gene names from a deliminated string}
\usage{
parseGeneNames(genes)
}
\arguments{
\item{genes}{an input string to be parsed}
}
\value{
a character vector with one element for every gene
}
\description{
Newline, tab, colon and semicolon and comma delimited lists are supported
}
\examples{
parseGeneNames("a b c")
parseGeneNames("a, b, c")
# etc.
}
