% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCombinedProfiles.r
\name{plotCombinedProfiles}
\alias{plotCombinedProfiles}
\title{plots combined profiles for a set of genes}
\usage{
plotCombinedProfiles(data, directions = F, colour = F, alpha = 0.2)
}
\arguments{
\item{data}{values returned by selectPlotData(profiles$counts)}

\item{directions}{if selected, the genes are split according to the direction of change in the primary lines}

\item{colour}{if true, the colours of the profiles are differentiated to help 'link' different genes by eye}

\item{alpha}{alpha value to be applied to each line}
}
\value{

}
\description{
Plots the combined profiles for a set of genes. Note that we need the extended four states on the horizontal axis as
cannot otherwise  link the separate plots for genes. The colour scheme (optional) is to help follow through individual
gene profiles
}
\examples{
data("profiles")
plotCombinedProfiles(selectPlotData(profiles$counts,gene=c("SLAMF8","DENND4A","VILL","NAB2","SRRM5")))
plotCombinedProfiles(selectPlotData(profiles$counts,gene=c("SLAMF8","DENND4A","VILL","NAB2","SRRM5")),colour=T)
plotCombinedProfiles(selectPlotData(profiles$counts,gene=c("SLAMF8","DENND4A","VILL","NAB2","SRRM5")),colour=T,directions=T)
}
