% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGeneDist.r
\name{plotGeneDist}
\alias{plotGeneDist}
\title{plots the distribution of the genes}
\usage{
plotGeneDist(
  x,
  type = "summary",
  log = F,
  guides = NULL,
  guides2 = NULL,
  wrap = F,
  no.guides = F
)
}
\arguments{
\item{x}{the measured gene data}

\item{type}{the type of plot required: currently only does a simple summary plot}

\item{guides}{if given, red reference lines are shown for the specified values}

\item{guides2}{if given, blue reference lines are shown for the specified values}

\item{wrap}{if true, separate distributions are plotted for each cell type. Redundant for
quantile normalised data where they are forced to be the same}

\item{no.guides}{if selected, all guide lines are suppressed}
}
\description{
plots the distribution of the genes
}
\examples{
plotGeneDist(newProfiles$counts)
plotGeneDist(newProfiles$counts,guides=c(10,16))
}
