% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotProfile.r
\name{plotProfile}
\alias{plotProfile}
\title{plots profile for a gene}
\usage{
plotProfile(
  data,
  multiscale = F,
  markSamples = NULL,
  log = F,
  id = T,
  size = 3
)
}
\arguments{
\item{data}{values returned by selectPlotData(profiles$values)}

\item{multiscale}{if true, use different scales for each plot (multiplot only)}

\item{markSamples}{a vector of patterns. If a sample is contains one of the patterns
given in the vector, it is identified as marked on the plot. If NULL, the default,
no action is taken.}

\item{log}{plot on a log scale}

\item{id}{plot sample IDs}

\item{size}{size of the plotting symbols}
}
\description{
Now handles multiple genes. The gene name is printed as a facet banner at the top.
}
\examples{
data("newProfiles")
plotProfile(selectPlotData(newProfiles$counts))
plotProfile(selectPlotData(newProfiles$counts,gene=c("SLAMF8","DENND4A","VILL","NAB2","SRRM5")))
plotProfile(selectPlotData(newProfiles$counts,gene=c("SLAMF8","DENND4A","VILL","NAB2","SRRM5")),multiscale = T)
}
