% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primaryDataDistribution.r
\name{primaryDataDistribution}
\alias{primaryDataDistribution}
\title{computes min and max gene values for primary}
\usage{
primaryDataDistribution(x)
}
\arguments{
\item{x}{the gene data}
}
\value{
a table of mins, maxes, min_maxes etc by gene (and state)
}
\description{
computes min and max gene values for primary
}
