% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectPlotData.r
\name{selectPlotData}
\alias{selectPlotData}
\title{selects profile data of interest for a given gene}
\usage{
selectPlotData(counts, gene = "MFSD14A")
}
\arguments{
\item{counts}{the counts component from the surplied profiles dataset}

\item{gene}{the requested gene of interest. MUltiple genes are allowed}
}
\value{

}
\description{
selects profile data of interest for a given gene
}
\examples{
data(profiles)
selectPlotData(profiles$counts)
}
