% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryForSelectedGenes.r
\name{summaryForSelectedGenes}
\alias{summaryForSelectedGenes}
\title{Title}
\usage{
summaryForSelectedGenes(genes, rv, contrasts = c("primary", "PMA"))
}
\arguments{
\item{contrasts}{}
}
\value{

}
\description{
Title
}
