% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textPage.r
\name{textPage}
\alias{textPage}
\title{plots one or more lines of text as a graphical plot}
\usage{
textPage(x = date())
}
\arguments{
\item{x}{the text to be plotted}
}
\value{

}
\description{
Plots a string of text as a graphical plot. This may be done as a
single line or multiple if new line characters are included. Lines
and text are horizontally and vertically centered and centre justified.
}
\examples{
textPage()
textPage("hello")
textPage("hello\nand\nwelcome")
}
