% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewer.r
\name{viewer}
\alias{viewer}
\title{top level function to run the app}
\usage{
viewer(source = "v2")
}
\arguments{
\item{the}{data source}
}
\description{
top level function to run the app
}
\examples{
viewer()
}
